Implements CSSPropertyAPI for the color properties.
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyWebkitTextFillColor, CSSPropertyWebkitTapHighlightColor,
CSSPropertyWebkitTextEmphasisColor, CSSPropertyWebkitBorderStartColor,
CSSPropertyWebkitBorderEndColor, CSSPropertyWebkitBorderBeforeColor,
CSSPropertyWebkitBorderAfterColor, CSSPropertyWebkitTextStrokeColor,
CSSPropertyStopColor, CSSPropertyFloodColor, CSSPropertyLightingColor,
and CSSPropertyColumnRuleColor from the switch statement in
parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
color properties is stored in a CSSPropertyDescriptor, and is called
from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIColor.cpp to the BUILD.gn file.
- Adds api_class flag to CSSProperties.in, which indicates that
  CSSPropertyAPIColor.h is generated.
- Moves the parsing logic for color from CSSPropertyParser.cpp to
  CSSPropertyAPIColor.cpp, which implements CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2647613002
Cr-Original-Commit-Position: refs/heads/master@{#445619}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: db1647d92fc600fae443c96fa5588271aa510267