Implements CSSPropertyAPI for the image properties.
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyListStyleImage, CSSPropertyBorderImageSource, and
CSSPropertyWebkitMaskBoxImageSource from the switch statement in
parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
image properties is stored in a CSSPropertyDescriptor, and is
called from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIImage.cpp to the BUILD.gn file.
- Adds api_class=CSSPropertyAPIImage flag to each radius class
  property in CSSProperties.in, which indicates that
  CSSPropertyAPIImage.h is generated.
- Moves the parsing logic for radius from CSSPropertyParser.cpp
  to CSSPropertyAPIImage.cpp, which implements CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2644023002
Cr-Original-Commit-Position: refs/heads/master@{#445621}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a59f5f57b230e4c0e42e41a907f274a53f718e1