Make logging in test_importer.py quieter.
Reason: Currently, files may be skipped when importing, but the reason
why they're skipped is buried in hundreds of lines of logging. This
CL changes logging so that all entries listing individual files and
directories in normal operation are changed to "debug" level.

BUG=684029

Review-Url: https://codereview.chromium.org/2654873003
Cr-Original-Commit-Position: refs/heads/master@{#445798}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 76cbd1cf8bb3c100994df4d28a8357008428ea8f