Implements CSSPropertyAPI for -webkit-max-logical-{width,height} and max-{inline,block}-size
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyMaxInlineSize, CSSPropertyMaxBlockSize,
CSSPropertyWebkitMaxLogicalWidth and CSSPropertyWebkitMaxLogicalHeight
from the switch statement in parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
affected properties is stored in a CSSPropertyDescriptor, and is
called from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIWebkitMaxLogicalWidthOrHeight.cpp to the BUILD.gn file.
- Adds api_class=CSSPropertyAPIWebkitMaxLogicalWidthOrHeight flag to each
  affected property in CSSProperties.in, which indicates that
  CSSPropertyAPIWebkitMaxLogicalWidthOrHeight.h is generated.
- Moves the parsing logic for max-inline-size, max-block-size,
  -webkit-max-logical-width and -webkit-max-logical-height from
  CSSPropertyParser.cpp to CSSPropertyAPIWebkitMaxLogicalWidthOrHeight.cpp,
  which implements CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2644703004
Cr-Original-Commit-Position: refs/heads/master@{#445980}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c5a725556b091a3db0970bb42c903ae08c6496a0