Rename w3c test importer script update-w3c-deps -> wpt-import.
Reason: This name is meant to be clearer and more consistent with related scripts.

Note that I want to call it "wpt-import" instead of "w3c-import" because there are plans to merge csswg-test into wpt, and I want to call it "wpt-import" instead of "wpt-update" because this makes its relationship with "wpt-export" clearer.

BUG=681406

Review-Url: https://codereview.chromium.org/2650923003
Cr-Original-Commit-Position: refs/heads/master@{#446044}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6fe839799893f3d82b5ab6ccfe0e759f8044c091