Migrate WTF::Vector::append() to ::push_back() [part 17 of N]
As discussed on platform-architecture-dev [0], there is interest in
normalizing the methods of WTF classes to better align with std classes.
This CL fixes a number of scattered references to append() that were
missed in previous CLs or added to the code since then.

There are no functional changes.

[0] https://groups.google.com/a/chromium.org/d/topic/platform-architecture-dev/I7jnz4p1h84/discussion

BUG=662431

Review-Url: https://codereview.chromium.org/2653203003
Cr-Original-Commit-Position: refs/heads/master@{#446074}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 54aad911b25fd13622a400c73fc8b3bd0026ebda