DrawingBuffer: Use gfx::GpuMemoryBuffer directly and set color space
With this instance remove, we no longer have any non-test callers of
CreateGpuMemoryBufferImageCHROMIUM.

Set the TextureMailbox color space for the DrawingBuffer. This
parameter is ignored except when --enable-color-correct-rendering is
used (further in the compositing pipeline).

Set the scanout color space for the DrawingBuffer, but only when
--enable-color-correct-rendering is used (because that function
goes directly to the underlying IOSurface).

BUG=648707
TBR=kbr

Review-Url: https://codereview.chromium.org/2643343002
Cr-Original-Commit-Position: refs/heads/master@{#446156}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b8cf426b59b931395bfe5114c7c6952473f210ce