Migrate WTF::HashSet::add() to ::insert() [part 2 of N]
As discussed on platform-architecture-dev [0], there is interest in
normalizing the methods of WTF classes to better align with std classes.
This CL fixes many (but not all) references to use insert() instead of
add() in subclasses of HashSet.

There are no functional changes.

[0] https://groups.google.com/a/chromium.org/d/topic/platform-architecture-dev/I7jnz4p1h84/discussion

BUG=662431

Review-Url: https://codereview.chromium.org/2654033007
Cr-Original-Commit-Position: refs/heads/master@{#446412}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fa8faf8ffd77d4629e3471372d3db338349fa931