PlzNavigate: Attempt to fix blink layout tests which fail due to duplicate output from WebFrameClient::willSendRequest.
With PlzNavigate, the willSendRequest function is called in RenderFrameImpl::BeginNavigation() and again during OnCommitNavigation()
when the request is loaded by blink. One option which was considered was to call willSendRequest() in OnCommitNavigation(). However the
request may be blocked and hence that call may never happen.

To workaround this I added a flag to track renderer initiated navigations to the RequestExtraData structure. We use this flag to determine
whether the WebFrameTestClient::willSendRequest() function needs to early return or not. The assumption is that the logging would already
be done when the navigation was initiated for renderer initiated navigations. No need to log when it is committed.

The other changes are to add functionality to the WebTestDelegate interface for the runner to query this flag.

There is one more test marked under the same bug http/tests/loading/redirect-methods.html. Does not appear to be the same issue. Will debug and log
a new bug if required.

BUG=625765

Review-Url: https://codereview.chromium.org/2652123002
Cr-Original-Commit-Position: refs/heads/master@{#446462}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fce540a08089df75d18cd4d719bec9ca03e88ac5