Rename RTCPeerConnection.updateIce to setConfiguration and make it work.
In addition to the renaming/basic wiring up, this CL adds a mechanism
for passing error information between webrtc/chromium and blink.

In this CL, it's only used for an InvalidModificationError.

Intent to Implement and Ship setConfiguration:
https://groups.google.com/a/chromium.org/d/msg/blink-dev/wh74Pd37LUQ/45Dx81Y2BgAJ

Intent to Deprecate and Remove updateIce:
https://groups.google.com/a/chromium.org/d/msg/blink-dev/9gq36Wl4Jwo/a8-rZOqCAgAJ

BUG=chromium:587453

Review-Url: https://codereview.chromium.org/2631433002
Cr-Original-Commit-Position: refs/heads/master@{#446512}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9853d6c5883e860ff52b5447b17f50907ffee434