Move webkit-patch bash completion to tools/bash-completion.
Most of the contents of webkit-tools-completion.sh is now irrelevant,
but there was some logic for webkit-patch command completion there.

This CL doesn't copy it over, but adds some basic completion for
webkit-patch to tools/bash-completion.

Thanks to lukasza@, there is already completion for run-webkit-tests
there (added in https://codereview.chromium.org/2435893002).

This also makes a minor change to bash-completion to make it work
for zsh as well (if $BASH_SOURCE is undefined, use $0 instead).

BUG=660890

Review-Url: https://codereview.chromium.org/2657943006
Cr-Original-Commit-Position: refs/heads/master@{#446551}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 85380c97c1bc176e8f40f0649cf21e419d036316