Show service worker navigation preload requests in DevTools Network tab.
Demo: https://youtu.be/I-Qe_Y-xYxE

Previously I tried to implement this feature by adding InspectorNetworkAgent in
the WorkerInspectorController of Service Worker.
https://codereview.chromium.org/2620463002/
But InspectorNetworkAgent is not designed to be used at worker threads. It
has introduced several DCHECK crashes. So I reverted it.

This CL sends the notifications to the DevTools from the browser process. This
design was proposed by pfeldman@.
https://codereview.chromium.org/2620463002/#msg115

BUG=649558

Review-Url: https://codereview.chromium.org/2654993006
Cr-Original-Commit-Position: refs/heads/master@{#446560}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6f9e15e126016658be28422100f9c4d1d4da6804