Paint invisible layer content in presence of composited animations.
For SPv2 we must paint layer content even if the content is invisible
w.r.t. opacity.

This is required (a) so that we have a paint chunk, which will produce a layer,
which is required by the animation subsystem, and (b) so that cc has the paint
commands needed in order to actually enact the animation compositor-side through
mutating the property tree state and re-rastering as needed.

In SPv1 this code path (PaintLayerPainter::paintedOutputInvisible) is skipped
when painting a layer for composited animated content. Instead of calling
PaintLayerPainter::paint(), we end up painting the GraphicsLayer for the
composited animation via CompositedLayerMapping::doPaintTask ->
PaintLayerPainter::paintLayerContents.

BUG=684842
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2652273004
Cr-Original-Commit-Position: refs/heads/master@{#446605}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9ce0839c605303284391c7efacccdcc00dc38437