Rename the WebUSB Origin Trial to "WebUSB2".
This rename has no effect on any of the feature flags but will require
that new Origin Trial keys for the feature be issued for "WebUSB2". This
ensures a clean break between the original trial period (Chrome 54-56)
and the new trial period (Chrome 57-59).

BUG=685756

Review-Url: https://codereview.chromium.org/2657443008
Cr-Original-Commit-Position: refs/heads/master@{#446854}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fb07160473f518862e5a8b5d3791266be5e5a91b