Implements CSSPropertyAPI for the alignment properties.
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyAlignItems, CSSPropertyAPIJustifyItems,
CSSPropertyAPIAlignSelf and CSSPropertyAPIJustifySelf from the
switch statement in parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
align-items, align-self, justify-items and justify-self properties is
stored in a CSSPropertyDescriptor, and is called from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIAlignItems.cpp, CSSPropertyAPIJustifyItems.cpp and
  CSSPropertyAPIAlignOrJustifySelf.cpp to the BUILD.gn file.
- Adds api_class flag to align-items and justify-items, and
  api_class=CSSPropertyAPIAlignOrJustifySelf to align-self and
  justify-self in CSSProperties.in, which indicates that all the .h
  files are generated.
- Moves the parsing logic for page from CSSPropertyParser.cpp to
  these .cpp files, which implement CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2640493004
Cr-Original-Commit-Position: refs/heads/master@{#447146}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2a9afa9fbfc5d76d668f3aeb70894f83d4686886