Replace bitwise & on EClear enum with equality.
Currently EClear is defined so that ClearLeft = 1, ClearRight = 2, and
ClearBoth = 3. Thus, we can use EClear like a bit flag and check if a
clear value X is either left or both by writing X & ClearLeft. However,
when we genereate EClear as part of generating ComputedStyle, there's
no guarantee on the ordering of the enum values, so we can no longer
depend on this trick. This patch simply replaces the bitwise operation
with an equality expression which doesn't depend on the ordering of
enum values. This is prework for converting EClear to an enum class.

BUG=665272

Review-Url: https://codereview.chromium.org/2652103002
Cr-Original-Commit-Position: refs/heads/master@{#447190}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b2ba6289f11d85440727609ac8430e7804933066