Clean up make_computed_style_base.py.
make_computed_style_base.py currently generates ComputedStyleBase, which
forms part of the ComputedStyle class. In preparation for generating
nonproperty fields in ComputedStyle, this patch refactors the script to
make it easier to add nonproperty fields and fixes a few issues with
the script, namely:
- Separate ComputedStyleBaseWriter.__init__ into smaller functions
- Rename property to property_ to avoid shadowing the property keyword
- Fixed a typo in asserting the expected vs real size
- Fixed an assert that ensures a field must have a valid field family

This patch does not change the generated ComputedStyleBase in any way.
This is prework for generating nonproperty fields, which would be
implemented as a new field family called 'nonproperty'.

BUG=628043

Review-Url: https://codereview.chromium.org/2661513002
Cr-Original-Commit-Position: refs/heads/master@{#447438}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1a59e8372660ac6452651b81a1900cecf0e7b7be