Move SetSinkIdResolver timer to frame-specific TaskRunnerTimer.
Move SetSinkIdResolver timer m_timer to frame-specific TaskRunnerTimer.
This associates it with the frame's MiscPlatformAPI timer task queue.
It is a speced task, so we should use MiscPlatformAPI.[1]

[1] The spec link: https://w3c.github.io/mediacapture-output/#htmlmediaelement-
extensions

This CL is migrating tasks/timers from a per-renderer scheduler to a per-frame
scheduler:
https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/80dsbEEOcWY

BUG=624694

Review-Url: https://codereview.chromium.org/2644153002
Cr-Original-Commit-Position: refs/heads/master@{#448184}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 89d445684f181b98567d6181670b6f84de2c981d