Convert GainNode tests to testharness
Manually converted the tests to use testharness.

Also modified gain.html not to create an audio file.  Instead, we load
up the reference output and compare against the output of the gain
node.  Test passes if the individual samples are close enough and the
SNR is high enough.

BUG=688169
TEST=gain-basic.html, gain.html

Review-Url: https://codereview.chromium.org/2671883002
Cr-Original-Commit-Position: refs/heads/master@{#448418}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 48b791c315ea56c1cd29a01283e466afd3dc4f08