Propagate fixed descendants of abs.
Bug: fixed descendants of absolute blocks were not getting positioned.
This fix was a TODO that fell off the radar.

Example:
<div id="rel" style="position:relative;">
  <div id="abs" style="position:absolute;top:100px;">
    <p>absolute</p>
    <div class="fix" style="position:fixed;background-color:gray;"><p>fixed</p></div>
  </div>
</div>

The loop inside NGOutOfFlowLayoutPart::Run had to be modified from
collection iterator to an array, because collections can't be modified
while iterating.

I've tried testing it in vertical-lr, but there are upstream bugs
that prevent this.

[ng_fixed_in_abs]

BUG=635619

Review-Url: https://codereview.chromium.org/2651393003
Cr-Original-Commit-Position: refs/heads/master@{#448463}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc062b30eac5feb37dcdc824cdf114b2aa30d77e