Remove replaceComposition() calls in finishComposingText.
We recently split commitText("") into a new method finishComposingText, a no-op
commit that simply preserves the already existing composition.  But because
of its historical origins it still had the replaceComposition call in it.

This is causing a bug with "yellow" autofill status getting cleared due
to the operation being interpreted as user typing.  But also, it's unnecessary work
in general, and the input/compositionupdate events are just repeated from the last
setComposition() call.  So just clean them out and have a clear()/moveCaret() call.

As a result of this, I was also able to clean out another longstanding autofill
ignoring flag that no longer has anything to ignore.

TEST=WebViewTest.FinishComposingTextDoesntTriggerAutofillTextChange
BUG=681822

Review-Url: https://codereview.chromium.org/2664603002
Cr-Original-Commit-Position: refs/heads/master@{#448546}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 81279cd5b209a34ca4f7702f2a10576faa50fdf9